Refactor: 온보딩 타입 포함 폴더 네이밍 변경
